+++ libs/common/src/lib/interfaces/portfolio-public-details.interface.ts
@@ -13,6 +13,7 @@ export interface PortfolioPublicDetails {
       | 'netPerformancePercent'
       | 'sectors'
       | 'symbol'
+      | 'url'
       | 'value'
     >;
   };