+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added an icon and name column to the positions table
+
 ### Changed
 
 - Changed the data gathering by symbol endpoint to delete data first